refactor: move MCP transport ownership to gateway
