+++ packages/core-mcp-server/src/index.ts
@@ -1,6 +1,4 @@
-export * from "./executor-session-manager.js";
 export * from "./mcp-router.js";
-export * from "./server.js";
 export * from "./tool-registry.js";
 export * from "./tool-snapshot.js";
 export type * from "./types.js";
